Agregado de edicion y creacion de stock

+++ TP3/Guglielmino.Jennifer.2E.TP3/Entidades/Interfaces/IVendibles.cs
@@ -8,7 +8,7 @@ namespace Entidades
 {
     interface IVendibles
     {
-        public float Precio { get; set; }
+        public double Precio { get; set; }
         bool Vender();
     }
 }
